Create actions to test and package things when we have new PRs
Summary

We now have a set of action to test the client when a new PRs is
created.
And also actions to publish a new release and push it to pypi when we
create a new version (ie. v1.0.0).

This is very similar to what we have done for `stacky` and it worked
well.

Testing

Let' see the test results