Fix cursor.py failing in SELECT * queries (#16)
There was a recent change in Rockset that removes the `column_fields`
field rather than keeping it an empty array in cases where columns are
not specified in the query (i.e. `SELECT *`), which causes an exception
to be raised when the field is accessed.

This change should resolve the crash and return the library to its
previous behavior.